move file to root dir
