+++ InternalSigner_test.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/bitcoinsv/bsvutil"
 
-	"github.com/libsv/libsv/bt"
-	"github.com/libsv/libsv/script"
+	"github.com/libsv/go-bt"
+	"github.com/libsv/go-bt/script"
 )
 
 func TestSignAuto(t *testing.T) {
