+++ Tx_test.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/bitcoinsv/bsvutil"
 
-	"github.com/libsv/libsv/bt"
-	"github.com/libsv/libsv/script"
+	"github.com/libsv/go-bt"
+	"github.com/libsv/go-bt/script"
 )
 
 func TestNew(t *testing.T) {
