+++ script/address/AddressValidation_test.go
@@ -3,7 +3,7 @@ package address_test
 import (
 	"testing"
 
-	"github.com/libsv/libsv/script/address"
+	"github.com/libsv/go-bt/script/address"
 )
 
 func TestValidateAddressMainnetP2PKH(t *testing.T) {
