+++ sighash/SigHashFlag.go
